+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Changed how relative lengths are computed in `offset`, `x` and `y`, now uses the maximum bounded length from constraint, the same as `size` properties.
 * Fix display print of `FactorPercent` not rounding.
 * `ChildInsert::{Over, Under}` now allows insert to affect layout size, like other inserts.
     - Use `background` and `foreground` properties as *layout passive* alternatives for `child_under` and `child_over`.